Extract speedfactor and framerate updates.
They now reside in their own method, which makes it a little easier for
overview. However, that method should rather be private, but due to
TSC’s larger number of mini mainloops (actions blocking all gameplay
except for a specific action) it needs to be public so these mini
mainloops can call it.

The mini mainloops should be removed, they’re an architectual flaw in
TSC. See the next commit for an example mini mainloop using the new
Update_Framerate() method.