Replace RenderWindow with RenderTarget for Draw().
SFML can render on other targets than only on the output window. While
this plays no role in normal gameplay (cScene::Draw() will always get
passed a sf::RenderWindow instance there), it matters for example when
taking screenshots. In that case, SFML must render onto another target
than the window, and this would not be possible if we restricted
cScene::Draw() to output to a visible window.

This commit establishes a necessary change for work on ticket #476.