Add GameOver scene background by screenshot.
The background of the GameOver scene is now created simply by taking a
screenshot of the current level, converting it into an sf::Texture, and
using that texture as the source of a new sf::Sprite which constitutes
the background of the GameOver scene. The “GameOver” text sprite is then
just drawn on top of this. This way the GameOver scene consists of only
two sprites, one could modify the background image (like greying it out
for example), and the mini mainloop that original OpenGL/SDL TSC had at
this point is prevented. And it is all nicely encapsulated into its own
scene, which it really is from a logical point of view.

People who work on #476 should take a look at how the screenshooting
functionality was implemented in level_player.cpp in this commit. Taking
real screenshots for saving them out to disk will function the same way,
just converting the sf::Texture back to an sf::Image again, which can be
saved out to a file.